fix: address PR review — 6 fixes
1. fix EXPON.DIST flag: remove incorrect RANGE_EXPAND
2. fix collect_from_node: use meta.name for inner match
3. widen FnCaps from u8 to u16 for future flags
4. make FunctionEntry::handler pub(crate)
5. exact entry count test (208)
6. aggregate-implies-NEEDS_PRELUDE invariant test

+++ crates/xlstream-parse/src/function_meta.rs
@@ -24,7 +24,7 @@ bitflags! {
     /// assert!(!caps.contains(FnCaps::LOOKUP));
     /// ```
     #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-    pub struct FnCaps: u8 {
+    pub struct FnCaps: u16 {
         /// Deterministic, no side effects.
         const PURE           = 0b0000_0001;
         /// Lazy arg evaluation (IF, AND, OR).
